requirements.txt
aenum==2.2.4
AnyQt==0.0.10
argon2-cffi==20.1.0
atpublic==2.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.2.1
confapp==1.1.11
cryptography==3.0
cycler==0.10.0
dask==2.22.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flake8==3.8.3
globus-sdk==1.9.0
graphviz==0.14.1
ibllib==1.5.5
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.4
jupyterlab-server==1.2.0
kiwisolver==1.2.0
libusb==1.0.22b9
llvmlite==0.33.0
logging-bootstrap==1.0.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mtscomp==1.0.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numba==0.50.1
numpy==1.19.1
opencv-python==3.4.5.20
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pge-plugin-terminal==0.1
phylib==2.2
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
pyarrow==1.0.0
pybpod==1.8.2
pybpod-api==1.8.2
pybpod-gui-api==1.8.2
pybpod-gui-plugin==1.8.2
pybpod-gui-plugin-alyx==1.1.2
pybpod-gui-plugin-emulator==0.1.5
pybpod-gui-plugin-rotaryencoder==0.1.4
pybpod-gui-plugin-session-history==1.4.2
pybpod-gui-plugin-soundcard==0.1.6
pybpod-gui-plugin-stmdiagram==1.0.0
pybpod-gui-plugin-timeline==1.0.1
pybpod-gui-plugin-trial-timeline==1.0.0
pybpod-gui-plugin-waveplayer==1.0
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pyflakes==2.2.0
pyforms-generic-editor==1.5.1
PyForms-GUI==4.901.2
pygame==1.9.6
Pygments==2.6.1
PyJWT==1.7.1
pynrrd==0.4.2
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQtWebEngine==5.15.0
pyrsistent==0.16.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
pyusb==1.0.2
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
QScintilla==2.11.5
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
safe-and-collaborative-architecture==0.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.10
visvis==1.12.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
